Fix ad loading while the user is scrolling
See also #167. Much like the Timer issue, when an NSURLConnection is created without specifying run loop modes, it defaults to `NSDefaultRunLoopMode`. Since this does not fire during `UITrackingRunLoopMode` it will not call the delegate, which prevents ads from loading.

This fix allows the delegates to fire during scrolling, which in turn allows ads to load.